pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edgevox"
version = "0.1.0"
description = "Sub-second local voice AI for robots and edge devices."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["voice-assistant", "robotics", "local-ai", "gemma", "whisper", "tts", "ros2", "edge-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=2.0",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "faster-whisper>=1.2",
    "llama-cpp-python>=0.3.8",
    "kokoro-onnx>=0.5",
    "onnxruntime>=1.21",
    "huggingface-hub>=0.24",
    "textual>=0.80",
    "rich>=13.0",
    "piper-tts>=1.2",
    "pymicro-wakeword>=2.2",
    "psutil>=5.9",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.21",
]
ros2 = ["rclpy"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "ruff>=0.8",
    "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
include = ["edgevox*"]

[tool.setuptools.package-data]
edgevox = ["py.typed"]

[project.scripts]
edgevox = "edgevox.tui:main"
edgevox-cli = "edgevox.cli.main:main"
edgevox-setup = "edgevox.setup_models:main"
edgevox-serve = "edgevox.server.main:main"

[project.urls]
Homepage = "https://github.com/edgevox-ai/edgevox"
Documentation = "https://edgevox-ai.github.io/edgevox/"
Issues = "https://github.com/edgevox-ai/edgevox/issues"

[tool.ruff]
target-version = "py310"
line-length = 120
extend-exclude = ["build", "dist", "website", "docs", "*.egg-info"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in default args (common in Textual/Click)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
